Fix YModem (binary stream) over the socket so file upload works
ioSender's macro upload to littlefs over YModem hung on the sim. Three socket
I/O bugs corrupted the binary stream (ASCII commands were unaffected, so this
went unnoticed):

1. NUL bytes dropped on receive. sim.getchar() returned 0 both for "byte 0x00"
   and for "no data", and mcu.c only forwarded non-zero values - so every 0x00
   in a YModem block (the 128-byte payload is mostly NUL padding) was silently
   dropped and the firmware never saw a complete block. Change the getchar
   contract to return int (0-255, or -1 for no data) across sim_socket_in /
   platform_poll_stdin / the sim_vars hook, and forward any byte >= 0.

2. Transmit line-buffering. sim_socket_out buffered output until a '\n'/'\r' or
   a full 128-byte buffer, but the YModem ACK/'C' handshake carries no newline,
   so the ACK sat unflushed and the host hung waiting for it. Send each byte as
   produced; the OS coalesces the small localhost writes.

3. 0x06 in the stream killed the sim. An incoming 0x06 (ACK / Ctrl-F) was always
   treated as a shutdown request, but it is a legal byte inside a binary stream.
   Only honour it in normal interactive mode, not during a YModem takeover.

Verified end to end against the sim: block 0 + data + EOT all ACK, and the file
lands on littlefs with the correct name, size and content ($F+ / $F<= dump).

+++ src/simulator.h
@@ -43,7 +43,7 @@ typedef struct sim_vars {
 #else
     int socket_fd;
 #endif
-    uint8_t (*getchar)(void);
+    int (*getchar)(void);   // returns 0-255 for a byte, or -1 when no data is available (NUL is a valid byte)
     void (*putchar)(uint8_t);
     sim_hook_fp on_init;
     sim_hook_fp on_tick;
